Relationship function update

+++ data-injestor/src/main/java/com/gline9/csci/hbase/Main.java
@@ -106,10 +106,10 @@ public class Main {
                         previousKey = metadataKey;
                         priceList.add(price);
                         System.out.println(overall + " and " +  count);
-                        overallList.add((double) (overall / count));
+                        overallList.add((double)overall / (double)count);
                         count = 1;
                         overall = Bytes.toShort(result.getValue(overallFamily, entry.getKey()));
-                        System.out.println("Adding price - " + price + " Overall - " + (double) (overall / count));
+                        System.out.println("Adding price - " + price + " Overall - " + (double)overall / (double)count);
                     }
                 }
             }
